1. Done minor changes related to db file path 2. Added condition for db existance or not 3. Removed unused code.

+++ Classifier.py
@@ -2,12 +2,12 @@ from database.table import get_all_records_count
 from database.table import get_all_match_records_count
 from database.table import get_all_weather_records
 from database.table import get_all_match_for_weather_count
-from enums.weather import WeatherCondition
-from enums.weather import Play
-from database.table import is_file_exists
 from database.table import create_database_connection
 from database.utils import delete_database
-from database.constant import db_name
+from database.utils import is_file_exists
+from database.utils import get_database_file_path
+from enums.weather import get_play_condition_string
+from enums.weather import get_weather_condition_string
 
 ''''
 Formula of naive bayes
@@ -21,32 +21,25 @@ P(E/H) - Likelihood of the evidence 'E' if the Hypothesis 'H' is true.
 '''
 
 
-# Problem : Find out the possibility of whether player play in Rainy condition ?
-# P(Yes|Rainy) = P(Rainy|Yes) * P(Yes) / P(Rainy)
-# P(Yes) - Find all yes from weather / All the weather records.
-# P(Rainy) - All rainy records / All the weather records.
-# P(Rainy/Yes) - All yes for rainy / Find all yes from weather
-
-
-def classify_play():
-    yes = get_all_match_records_count(Play.yes.value) / get_all_records_count()
-    rainy = get_all_weather_records(WeatherCondition.rainy.value) / get_all_records_count()
-    ry = get_all_match_for_weather_count(Play.yes.value, WeatherCondition.rainy.value) / \
-         get_all_match_records_count(Play.yes.value)
-    print('yes' + str(yes))
-    print('rainy' + str(rainy))
-    print('yes/rainy' + str(ry))
+def classify_play(weather_condition, play):
+    yes = get_all_match_records_count(play) / get_all_records_count()
+    rainy = get_all_weather_records(weather_condition) / get_all_records_count()
+    ry = get_all_match_for_weather_count(play, weather_condition) / \
+         get_all_match_records_count(play)
+    print('yes: ' + str(yes))
+    print(str(get_weather_condition_string(weather_condition)) + ': ' + str(rainy))
+    print('yes/' + str(get_play_condition_string(play)) + ': ' + str(ry))
     return (ry * yes) / rainy
 
 
-'''
-Create database connection before modeling
-'''
-
-
+# Creating database file name weather_report.db
 def create_database():
     try:
-        delete_database()
-        create_database_connection(db_name)
+        file_name = get_database_file_path()
+        # Check whether existing db file is available or not, if availbale deleted the old one.
+        if is_file_exists(file_name):
+            print("Deleted the existing db file")
+            delete_database()
+        create_database_connection(file_name)
     except:
-        create_database_connection(db_name)
+        print("Exception while creating a database")
